+++ UpgradeWorld/UpgradeWorld.cs
@@ -11,7 +11,7 @@ public class UpgradeWorld : BaseUnityPlugin
 {
   const string GUID = "upgrade_world";
   const string NAME = "Upgrade World";
-  const string VERSION = "1.74.8";
+  const string VERSION = "1.75";
 #nullable disable
   public static ManualLogSource Log;
 #nullable enable